+++ R/gt_fmt_cfb.R
@@ -19,7 +19,7 @@
 #'
 #'df <- data.frame(team = valid_team_names()[1:8],logo = valid_team_names()[1:8])
 #'
-#'df %>%
+#'table <- df %>%
 #'  gt() %>%
 #'  gt_fmt_cfb(columns = "logo")
 #'
@@ -31,10 +31,13 @@
 gt_fmt_cfb <- function(gt_object, columns, height = 30){
 
   # convert tidyeval column to bare string
-  col_bare <- rlang::enexpr(columns) %>% rlang::as_string()
+  column_names <- gt:::resolve_cols_c(
+    expr = {{ columns }},
+    data = gt_object
+  )
 
-
-  grp_var <- gt_object[["_boxhead"]][["var"]][which(gt_object[["_boxhead"]][["type"]]=="stub")]
+  stub_var <- gt_object[["_boxhead"]][["var"]][which(gt_object[["_boxhead"]][["type"]]=="stub")]
+  grp_var <- gt_object[["_boxhead"]][["var"]][which(gt_object[["_boxhead"]][["type"]]=="row_group")]
 
   # stopifnot("img_source must be 'web' or 'local'" = img_source %in% c("web", "local"))
   img_source <- "web"
@@ -42,8 +45,10 @@ gt_fmt_cfb <- function(gt_object, columns, height = 30){
   # need to correct for rownames
   gt_object %>%
     text_transform(
-      locations = if(isTRUE(grp_var == col_bare)){
-        cells_stub()
+      locations = if(isTRUE(grp_var %in% column_names)){
+        cells_row_groups()
+      } else if(isTRUE(stub_var %in% column_names)){
+        cells_stub(rows = gt::everything())
       } else {
         cells_body({{ columns }})
       },