+++ NEWS.md
@@ -4,3 +4,4 @@
 * Added the `geom_mean_lines()` and `geom_median_lines()` geoms.
 * Added the `scale_color_cfb()` and `scale_fill_cfb()` functions.
 * Added FCS team logos
+* added the `gt_fmt_cfb()` function.